% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFN.R
\name{showSinglePicture}
\alias{showSinglePicture}
\alias{showSingleColorPicture}
\alias{showPictures}
\title{Displays a picture, such as an MNIST digit or CIFAR image
Assumes the image is stored in a rowwise-fashion in the vector x}
\usage{
showSinglePicture(x, sidelength = NULL, col = gray(256:1/256), ...)

showSingleColorPicture(x, sidelength = NULL, normalize = TRUE, ...)

showPictures(data, nrow, ncol, idx = NULL, isColor = FALSE, ...)
}
\arguments{
\item{x}{the data to depict, a grayscale channel in \code{showSinglePicture}, or in
\code{showSingleColorPicture} it is assumed to be a vector where the first
third of the elements describe the red color channel, the next ones
the green channel and the last third the blue channel; in \code{showPictures} its interpretation
is controlled by the parameter \code{isColor}.}

\item{sidelength}{the height of the resulting picure in pixels; on \code{NULL}, a square picture
is assumed}

\item{col}{as passed to function \code{image}, see \code{?image}}

\item{...}{as passed to function \code{image}, see \code{?image}}

\item{normalize}{logical, whether \code{x} should be normalized}

\item{data}{contains the pixel data, one picture in each column}

\item{nrow, ncol}{define the arrangement of the pictures}

\item{idx}{defines the enumeration of the pictures, i.e. the order in which the columns of
\code{data} should be processed; on \code{NULL}, 1,2,3... is assumed}

\item{isColor}{logical, indicates whether \code{data} is to be interpreted as single channel
or 3-way-channel (RGB)}
}
\value{
\code{NULL}
}
\description{
Displays a picture, such as an MNIST digit or CIFAR image
Assumes the image is stored in a rowwise-fashion in the vector x

Assumes the image is stored in a rowwise-fashion.

Shows multiple colored or grayscale pictures
}
