% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFN.R
\name{somatcopy}
\alias{somatcopy}
\alias{gemm}
\title{for testing MKL}
\usage{
somatcopy(A)

gemm(A, B)
}
\arguments{
\item{A}{a matrix}

\item{B}{a matrix}

\item{A}{a sparse matrix}
}
\description{
for testing MKL

for testing MKL
}
