% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFN.R
\name{train_rfn}
\alias{train_rfn}
\title{Trains a Rectified Factor Network (RFN)}
\usage{
train_rfn(X, n_hidden, n_iter, etaW, etaP, minP, batch_size = -1,
  dropout_rate = 0, input_noise_rate = 0, l2_weightdecay = 0,
  l1_weightdecay = 0, h_threshold = 0, momentum = 0,
  noise_type = "saltpepper", activation = "relu", apply_scaling = 1,
  apply_newton_update = 1, seed = -1, use_gpu = 0, gpu_id = -1,
  verbose = T)
}
\arguments{
\item{X}{The data matrix}

\item{n_hidden}{Number of latent variables to estimate}

\item{n_iter}{Number of iterations to run the algorithm}

\item{etaW}{Learning rate of the W parameter}

\item{etaP}{Learning rate of the Psi parameter (It's probably save to set this to the same value as etaW)}

\item{minP}{Minimal value for Psi. Should be in 1e-5 - 1e-1}

\item{batch_size}{If > 2, this will activate mini-batch learning instead of full batch learning}

\item{dropout_rate}{Dropout rate for the latent variables}

\item{input_noise_rate}{Noise/dropout rate for input variables}

\item{l2_weightdecay}{L2 penalty for weight decay}

\item{l1_weightdecay}{L1 penalty for weight decay}

\item{h_threshold}{Threshhold for rectifying/leaky activations}

\item{momentum}{Momentum term for learning}

\item{noise_type}{Type of input noise. One of "dropout", "saltpepper", "gaussian"}

\item{activation}{Activation function for hidden/latent variables. One of "linear", "relu", "leaky", "sigmoid", "tanh"}

\item{apply_scaling}{Scale the data}

\item{apply_newton_update}{Whether to use a Newton update (default) or a gradient descent step}

\item{seed}{Seed for the random number generator}

\item{use_gpu}{Use the gpu (default cpu). Works only for sparse input.}

\item{gpu_id}{If use_gpu is true, use gpu with this id (default -1 selects one available)}

\item{verbose}{True to print messages during the run. Default is True.}
}
\value{
Returns a list of matrices \code{W}, \code{P}, \code{H},
 \code{Wout}, whereas \code{W \%*\% H} is the noise-free reconstruction
 of the data \code{X} and \code{diag(P)} is the covariance matrix
 of the additive zero-mean noise. \code{Wout} is the matrix that
 maps input vectors to their latent representation, usually by
 \code{pmax(t(Wout) \%*\% X, 0)}.
}
\description{
Trains an RFN as described by Clevert et al., 2014
}
